[REF] Move certifier.organization from product_food to product_food_certification [REF] product_food : move organic_type into product_food_certification [UPD] French translation [REM] Remove obsolete mass editing button for is_alimentary field.
