+++ product_food/__manifest__.py
@@ -9,30 +9,23 @@
     "author": "GRAP",
     "website": "https://github.com/grap/grap-odoo-business",
     "license": "AGPL-3",
-    "depends": ["product_label", "product_net_weight", "product_origin"],
+    "depends": ["product_label"],
     "data": [
         "security/res_groups.xml",
         "security/ir_rule.xml",
         "security/ir.model.access.csv",
-        "views/view_certifier_organization.xml",
         "views/view_product_label.xml",
         "views/view_product_allergen.xml",
         "views/view_product_category.xml",
         "views/view_product_product.xml",
         "views/view_product_template.xml",
-        "views/view_res_company.xml",
     ],
     "demo": [
-        "demo/res_groups.xml",
-        "demo/certifier_organization.xml",
         "demo/product_label.xml",
         "demo/product_allergen.xml",
         "demo/product_category.xml",
         "demo/product_product.xml",
     ],
-    "images": [
-        "static/description/certifier_organization_form.png",
-        "static/description/product_allergen_form.png",
-    ],
+    "external_dependencies": {"python": ["unidecode", "bs4"]},
     "installable": True,
 }
