+++ product_food/models/__init__.py
@@ -1,7 +1,5 @@
-from . import certifier_organization
 from . import product_allergen
 from . import product_label
 from . import product_category
 from . import product_product
 from . import product_template
-from . import res_company
