+++ product_food/readme/CONFIGURE.rst
@@ -2,16 +2,14 @@
 
 .. figure:: ../static/description/product_allergen_form.png
 
-* Go to 'Sale > Configuration > Certifier Organizations' and create new items.
-
-.. figure:: ../static/description/certifier_organization_form.png
-
-* Go to 'Inventory > Configuration > Product Categories' and check new
-  field ``is_alimentary`` for categories that contains food products.
+* Go to 'Inventory > Configuration > Product Categories' and check new fields:
+    * ``Contain Alimentary Products``
+    * ``Contain Vegan Products``
+    * ``Contain Alcohol Products``
 
 .. figure:: ../static/description/product_category_form.png
 
 * Go to 'Sale > Configuration > Product Labels' and update your labels,
-  setting two new fields ``organic_type`` and ``is_alcohol``.
+  setting two new fields ``Has Alcohol`` and ``Is Vegan``.
 
 .. figure:: ../static/description/product_label_form.png
