+++ product_food/readme/USAGE.rst
@@ -1,6 +1,3 @@
-To use this module, you need to to your product variants or templates
-and set the information in the following two tabs.
+* Go to your product variants or templates form and set the information in the following two tabs.
 
-.. figure:: ../static/description/product_label_form_alimentary_information.png
-
-.. figure:: ../static/description/product_label_form_alimentary_composition.png
+.. figure:: ../static/description/product_product_alimentary_tab.png
