+++ product_food/security/ir.model.access.csv
@@ -1,5 +1,3 @@
 id,name,model_id:id,group_id:id,perm_read,perm_write,perm_create,perm_unlink
-access_certifier_organization_manager,Certifier Organization Manager,model_certifier_organization,group_certifier_manager,1,1,1,1
-access_certifier_organization_all,Certifier Organization All,model_certifier_organization,,1,,,
 access_product_allergen_manager,Product Allergen Manager,model_product_allergen,group_allergen_manager,1,1,1,1
 access_product_allergen_all,Product Allergen All,model_product_allergen,,1,,,
