+++ product_food/security/res_groups.xml
@@ -7,11 +7,6 @@ License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 -->
 <odoo>
 
-    <record id="group_certifier_manager" model="res.groups">
-        <field name="name">Certifier Organizations Creation</field>
-        <field name="category_id" ref="base.module_category_usability"/>
-    </record>
-
     <record id="group_allergen_manager" model="res.groups">
         <field name="name">Allergens Creation</field>
         <field name="category_id" ref="base.module_category_usability"/>
