+++ product_food/views/view_product_template.xml
@@ -27,8 +27,6 @@ License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
                         <field name="use_by_date_day" attrs="{'invisible': [('is_alimentary', '=', False)]}"/>
                         <field name="best_before_date_day" attrs="{'invisible': [('is_alimentary', '=', False)]}"/>
-                        <field name="organic_type"/>
-                        <field name="is_uncertifiable" attrs="{'invisible': [('is_alimentary', '=', False)]}"/>
                     </group>
                     <group string="Ingredients" attrs="{'invisible': [('is_alimentary', '=', False)]}" colspan="4">
                         <field name="ingredients" widget="html" nolabel="1"/>
