+++ product_food_certification/readme/CONFIGURE.rst
@@ -1,0 +1,3 @@
+* Go to 'Sale > Configuration > Certifier Organizations' and create new items.
+
+.. figure:: ../static/description/certifier_organization_form.png
