+++ product_food_certification/readme/ROADMAP.rst
@@ -1,4 +1,4 @@
 * make the field ``ingredient_origin_type`` mandatory if the label 'Euroleaf'
-  is used, in compliance with current legislation.
+  is used and the unit of the product is weighable, in compliance with current legislation.
 
 * Remove the post_init_hook, onced installed on grap / caap instances.
