+++ product_food_certification/security/res_groups.xml
@@ -1,4 +1,9 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <odoo>
 
+    <record id="group_certifier_manager" model="res.groups">
+        <field name="name">Certifier Organizations Creation</field>
+        <field name="category_id" ref="base.module_category_usability"/>
+    </record>
+
 </odoo>
