+++ product_food_certification/views/view_product_label.xml
@@ -9,11 +9,10 @@ License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
     <record id="view_product_label_form" model="ir.ui.view">
         <field name="model">product.label</field>
-        <field name="inherit_id" ref="product_label.view_product_label_form" />
+        <field name="inherit_id" ref="product_food.view_product_label_form" />
         <field name="arch" type="xml">
-            <field name="company_id" position="after">
-                <field name="is_alcohol"/>
-                <field name="is_vegan"/>
+            <field name="is_alcohol" position="before">
+                <field name="organic_type"/>
             </field>
         </field>
     </record>
