+++ requirements.txt
@@ -1,2 +1,4 @@
 # generated from manifests external_dependencies
+bs4
 openupgradelib
+unidecode
